Move quantization to Measure object (#180)
