+++ src/looper/sample.rs
@@ -1,15 +1,10 @@
 use sdl2::render::Renderer;
 
 use midi;
-use midi::{AbsMidiEvent, TypedMidiMessage, Note, MidiSink};
+use midi::{AbsMidiEvent, Note, MidiSink};
 use measure::*;
 use rustc_serialize::{Decodable, Encodable, Encoder, Decoder};
 
-#[derive(RustcDecodable, RustcEncodable, Debug, PartialEq, Eq)]
-pub struct QuantMidiEvent {
-    pub message: TypedMidiMessage,
-    pub quant: Quant,
-}
 
 pub struct Sample {
     // FIXME(#153): Improve performance of the event look up in sample
@@ -89,14 +84,7 @@ impl Sample {
 
     pub fn new(buffer: &[AbsMidiEvent], measure: &Measure, measure_shift: u32) -> Sample {
         let amount_of_measures = measure.amount_of_measures_in_buffer(buffer);
-        let quants_per_sample = Quant(amount_of_measures) * measure.quants_per_measure();
-
-        let quant_buffer: Vec<QuantMidiEvent> = buffer.iter().map(|event| {
-            QuantMidiEvent {
-                message: event.message,
-                quant: measure.snap_timestamp_to_quant(event.timestamp) % quants_per_sample,
-            }
-        }).collect();
+        let quant_buffer = measure.quantize_buffer(buffer);
 
         let notes = midi::events_to_notes(&quant_buffer);
