+++ src/measure/quant_midi_event.rs
@@ -1,0 +1,8 @@
+use midi::TypedMidiMessage;
+use measure::Quant;
+
+#[derive(RustcDecodable, RustcEncodable, Debug, PartialEq, Eq)]
+pub struct QuantMidiEvent {
+    pub message: TypedMidiMessage,
+    pub quant: Quant,
+}
